% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_specs.R
\docType{data}
\name{study_specs}
\alias{study_specs}
\title{Study-specific paths and settings}
\format{
An object of class \code{list} of length 1.
}
\usage{
study_specs
}
\description{
Study-specific paths and settings
}
\keyword{datasets}
